Move non view functions to utils.py
